Function factory for likelihood evaluation not using sparsity

Function factory for likelihood evaluation not using sparsity

Usage

likelihood_graph_covariance(
  graph,
  model = "alpha1",
  y_graph,
  cov_function = NULL,
  X_cov = NULL,
  repl,
  log_scale = TRUE,
  maximize = FALSE
)

Arguments

graph

metric_graph object

model

Type of model: "alpha1" gives SPDE with alpha=1, "GL1" gives the model based on the graph Laplacian with smoothness 1, "GL2" gives the model based on the graph Laplacian with smoothness 2, and "isoCov" gives a model with isotropic covariance.

y_graph

Response vector given in the same order as the internal locations from the graph.

cov_function

The covariance function to be used in case 'model' is chosen as 'isoCov'. cov_function must be a function of (h, theta_cov), where h is a vector, or matrix, containing the distances to evaluate the covariance function at, and theta_cov is the vector of parameters of the covariance function cov_function.

X_cov

Matrix with covariates. The order must be the same as the internal order from the graph.

repl

Vector with the replicates to be considered. If NULL all replicates will be considered.

log_scale

Should the parameters theta of the returning function be given in log scale?

maximize

If FALSE the function will return minus the likelihood, so one can directly apply it to the optim function.

Value

The log-likelihood function, which is returned as a function with parameter 'theta'. For models 'alpha1', 'alpha2', 'GL1' and 'GL2', the parameter theta must be supplied as the vector c(sigma_e, sigma, kappa).

For 'isoCov' model, theta must be a vector such that theta[1] is sigma.e and the vector theta[2:(q+1)] is the input of cov_function, where q is the number of parameters of the covariance function.

If covariates is TRUE, then the parameter theta must be supplied as the vector c(sigma_e, theta[2], ..., theta[q+1], beta[1], ..., beta[p]), where beta[1],...,beta[p] are the coefficients and p is the number of covariates.

For the remaining models, if covariates is TRUE, then theta must be supplied as the vector c(sigma_e, sigma, kappa, beta[1], ..., beta[p]), where beta[1],...,beta[p] are the coefficients and p is the number of covariates.